// src/text/enums.ts
// Fichier central pour tous les enums - évite les imports circulaires

export enum RegionType {
  Region = "regions",
  Ranch = "ranch",
}

export enum Region {
  Fields = "fields",
  Strand = "strand",
  Valley = "valley",
  Bluffs = "bluffs",
  Labyrinth = "labyrinth",
  Sea = "sea",
}

export enum Ranch {
  Conservatory = "conservatory",
  Archway = "archway",
  Den = "den",
  Gully = "gully",
  Tidepools = "tidepools",
  Digsite = "digsite",
}

export enum Weather {
  Clear = "clear",
  Rain = "rain",
  Snow = "snow",
  Thunder = "thunder",
  Lightning = "lightning",
  Wind = "wind",
  Cyclone = "cyclone",
  Pollen = "pollen",
  Vine = "vine",
  Slimefall = "slimefall",
}

export enum Resource {
  Primordy = "primordy",
  Brine = "brine",
  Sand = "sand",
  Lava = "lava",
  Wax = "wax",
  Honey = "wildhoney",
  Jelly = "jelly",
  Radiant = "radiant",
  Diamond = "diamond",
  Fossil = "fossil",
  Snowflake = "snowflake",
  Sunsap = "sunsap",
  Mote = "mote",
  Glass = "glass",
  Drift = "drift",
  Magma = "magma",
  Aqua = "aqua",
  Dream = "dream",
  Petal = "petal",
  Indigo = "blackindigo",
  Royal = "royal",
  Snowball = "snowball",
  Prisma = "prisma",
}